requirements.txt
flask
python-dotenv
gunicorn
psycopg2